refactor(react): simplify richtext renderer API
- remove hook factory abstraction
- rename files to follow standard naming conventions
- simplify StoryblokRichText component
- streamline RSC and client-side rendering of RichText
