+++ packages/react/src/richtext.ts
@@ -1,6 +1,15 @@
-import StoryblokComponent from './core/storyblok-component';
-import { createRichTextHook } from './core/richtext-hoc';
+import { createRichTextRenderer } from './core/richtext';
+import type { StoryblokRichTextProps } from './core/richtext';
+import { DefaultBlok } from './storyblok-rich-text';
 
-export const useStoryblokRichText = createRichTextHook(StoryblokComponent, {
-  isServerContext: false,
-});
+export const useStoryblokRichText = (props: StoryblokRichTextProps) => {
+  const render = createRichTextRenderer({
+    optimizeImage: props.optimizeImage,
+    components: {
+      blok: DefaultBlok,
+      ...props.components,
+    },
+  });
+
+  return render(props.document);
+};
