+++ packages/react/src/rsc/index.ts
@@ -1,10 +1,9 @@
 // React Server Components (includes server actions for live editing)
 export * from '../core/init';
 export * from '../core/richtext';
-export { useStoryblokServerRichText } from '../server/richtext';
 export { default as StoryblokServerComponent } from '../server/server-component';
 export { default as StoryblokServerStory } from '../server/server-story';
-export { default as StoryblokServerRichText } from '../server/server-storyblok-richtext-component';
+export { StoryblokServerRichText } from '../server/server-storyblok-richtext';
 export * from '../types';
 
 export { default as StoryblokLiveEditing } from './live-editing';
