+++ packages/react/src/server/richtext.tsx
@@ -1,6 +1,15 @@
-import StoryblokServerComponent from './server-component';
-import { createRichTextHook } from '../core/richtext-hoc';
+import { createRichTextRenderer } from '../core/richtext';
+import type { StoryblokRichTextProps } from '../core/richtext';
+import { DefaultBlok } from './server-storyblok-richtext';
 
-export const useStoryblokServerRichText = createRichTextHook(StoryblokServerComponent, {
-  isServerContext: true,
-});
+export const useStoryblokServerRichText = (props: StoryblokRichTextProps) => {
+  const render = createRichTextRenderer({
+    optimizeImage: props.optimizeImage,
+    components: {
+      blok: DefaultBlok,
+      ...props.components,
+    },
+  });
+
+  return render(props.document);
+};
