+++ packages/react/src/ssr/index.ts
@@ -1,8 +1,7 @@
 // SSR-compatible components (static-safe, no server actions)
 export * from '../core';
 export * from '../core/richtext';
-export { useStoryblokServerRichText } from '../server/richtext';
 export { default as StoryblokServerComponent } from '../server/server-component';
 
 export { default as StoryblokServerStory } from '../server/server-story';
-export { default as StoryblokServerRichText } from '../server/server-storyblok-richtext-component';
+export { StoryblokServerRichText } from '../server/server-storyblok-richtext';
